// src/app/models/excelObject.ts
export class excelObject {
    ResponseId: number;
    UserName: string;
    UserEmail: string;
    Question_1_Answer: string;
    Question_2_Answer: string;
    Question_3_Answer: string;
    Question_4_Answer: string;
    Question_5_Answer: string;
    Question_6_Answer: string;
    ResponseTime: string;
}